[ios] Disable implicit animation of heading indicators
The update steps for the heading indicator are typically small, so animations tend to pile up and cause performance issues. Disabling actions is a slight regression when it comes to large steps (they're not animated now, where they previously were) and this should eventually be addressed.

Also consistently use provided API for disabling CATransaction actions.